static/static/static/static/static/static/static/static/static/static/static/static/static/static/static/static/static/static/static/static/static/static/static/static/static/static/static/static/static/static/static/requirements.txt
asgiref==3.9.1
diff-match-patch==20241021
Django==5.2.4
django-import-export==4.3.8
numpy==2.3.1
pandas==2.3.1
pillow==11.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
